send 8x decimated freq noise data  (after noise projection) to logger
Warning: we are doing some maths on the frequency noise data before rounding to 16 bits to send to the logger. This creates an artificial white frequency noise floor around -20 dB(Hz^2/Hz) noise floor. This noise floor is only present on the logged data and does not apply to the PLL.